practiceautomatedtesting/shopping/requirements.txt
selenium>=4.15.2
pytest>=8.0.0
webdriver-manager>=4.0.1
allure-pytest>=2.13.2 